Move dashboard tests to tests/controllers/admin
